+++ auction-server/api-types/src/opportunity.rs
@@ -349,6 +349,10 @@ pub enum OpportunityParamsV1ProgramSvm {
         #[schema(example = "memo")]
         memo: Option<String>,
 
+        /// If true, bids to this opportunity can be cancelled by the searcher.
+        #[schema(example = true)]
+        cancellable: bool,
+
         /// The quote deadline (in seconds) must be at least this many seconds later than the time the request is received by the server.
         /// A value like `10` means the deadline must be at least 10 seconds after the request is received.
         /// To ensure the deadline is reasonable, take into account possible network latency, auction and submission delays.