+++ auction-server/src/opportunity/entities/opportunity_svm.rs
@@ -107,6 +107,7 @@ pub struct OpportunitySvmProgramSwap {
     pub token_program_searcher:               Pubkey,
     pub token_account_initialization_configs: TokenAccountInitializationConfigs,
     pub memo:                                 Option<String>,
+    pub cancellable:                          bool,
     pub minimum_lifetime:                     Option<u32>,
 }
 
@@ -177,6 +178,7 @@ impl Opportunity for OpportunitySvm {
                             .token_account_initialization_configs,
                         user_mint_user_balance:               program.user_mint_user_balance,
                         memo:                                 program.memo,
+                        cancellable:                          program.cancellable,
                         minimum_lifetime:                     program.minimum_lifetime,
                     },
                 )
@@ -407,6 +409,7 @@ impl From<OpportunitySvm> for api::OpportunitySvm {
                         .token_account_initialization_configs
                         .into(),
                     memo: program.memo,
+                    cancellable: program.cancellable,
                     minimum_lifetime: program
                         .minimum_lifetime
                         .unwrap_or(DEFAULT_SWAP_BID_MINIMUM_LIFE_TIME),
@@ -466,6 +469,7 @@ impl TryFrom<repository::Opportunity<repository::OpportunityMetadataSvm>> for Op
                         .token_account_initialization_configs,
                     user_mint_user_balance:               program.user_mint_user_balance,
                     memo:                                 program.memo,
+                    cancellable:                          program.cancellable,
                     minimum_lifetime:                     program.minimum_lifetime,
                 })
             }