+++ sdk/js/src/index.ts
@@ -520,11 +520,13 @@ export class Client {
         ? quoteRequest.userWallet.toBase58()
         : null,
       memo: quoteRequest.memo ?? null,
+      cancellable: quoteRequest.cancellable ?? undefined,
+      minimum_lifetime: quoteRequest.minimumLifetime ?? null,
       version: "v1" as const,
     };
     // TODO: we may want to wrap all the GET/POST calls in a try/catch block to handle errors
     const response = await client.POST("/v1/opportunities/quote", {
-      body: body,
+      body,
     });
     if (response.error) {
       throw ClientError.newHttpError(