+++ sdk/js/src/svm.ts
@@ -251,6 +251,7 @@ function extractSwapInfo(swapOpportunity: OpportunitySvmSwap): {
   tokenProgramSearcher: PublicKey;
   tokenInitializationConfigs: TokenAccountInitializationConfigs;
   memo?: string;
+  cancellable: boolean;
   userMintUserBalance: anchor.BN;
 } {
   const tokenProgramSearcher = swapOpportunity.tokens.tokenProgramSearcher;
@@ -265,6 +266,7 @@ function extractSwapInfo(swapOpportunity: OpportunitySvmSwap): {
   const router = swapOpportunity.routerAccount;
   const tokenInitializationConfigs = swapOpportunity.tokenInitializationConfigs;
   const memo = swapOpportunity.memo;
+  const cancellable = swapOpportunity.cancellable;
   const userMintUserBalance = swapOpportunity.userMintUserBalance;
   return {
     searcherToken,
@@ -277,6 +279,7 @@ function extractSwapInfo(swapOpportunity: OpportunitySvmSwap): {
     router,
     tokenInitializationConfigs,
     memo,
+    cancellable,
     userMintUserBalance,
   };
 }
@@ -468,6 +471,8 @@ export async function constructSwapBid(
     tokenInitializationConfigs,
     userMintUserBalance,
   } = extractSwapInfo(swapOpportunity);
+  const cancellable = swapOpportunity.cancellable;
+  console.log("cancellable", cancellable);
 
   if (swapOpportunity.memo) {
     tx.instructions.push(createMemoInstruction(swapOpportunity.memo));